CategoryDetail renamed to CategoryWithPreloadedChildren to make the purpose of the class clear
- CategoryDetailFactory::createDetailsHierarchy() moved to new CategoryWithPreloadedChildrenFactory::createCategoriesWithPreloadedChildren()
- CategoryFacade::getAllCategoryDetails() renamed to getAllCategoriesWithPreloadedChildren()
- CategoryFacade::getVisibleCategoryDetailsForDomain() renamed to getVisibleCategoriesWithPreloadedChildrenForDomain()
- twig macros for rendering category tree in admin use instances of CategoryWithPreloadedChildren class
