+++ packages/framework/src/Resources/views/Admin/Content/Category/list.html.twig
@@ -13,21 +13,21 @@
 {% block main_content %}
     {{ render(controller('ShopsysFrameworkBundle:Admin/Category:listDomainTabs')) }}
 
-    {% macro categoryTreeItem(categoryDetails, isFirstLevel) %}
+    {% macro categoryTreeItem(categoriesWithPreloadedChildren, isFirstLevel) %}
         {% import _self as self %}
         <ul class="js-category-tree-items {{ isFirstLevel ? 'form-tree__content' : '' }}">
-            {% for categoryDetail in categoryDetails %}
-                <li class="js-category-tree-item form-tree__item" id="js-category-tree-{{ categoryDetail.category.id }}">
+            {% for categoryWithPreloadedChildren in categoriesWithPreloadedChildren %}
+                <li class="js-category-tree-item form-tree__item" id="js-category-tree-{{ categoryWithPreloadedChildren.category.id }}">
                     <div class="js-category-tree-item-line form-tree__item__drag">
                         <span class="form-tree__item__icon form-tree__item__icon--level sprite sprite-level"></span>
-                        <span class="js-category-tree-item-line  js-category-tree-item-handle form-tree__item__name">{{ categoryDetail.category.name }}</span>
+                        <span class="js-category-tree-item-line  js-category-tree-item-handle form-tree__item__name">{{ categoryWithPreloadedChildren.category.name }}</span>
                         <span class="js-category-tree-item-icons form-tree__item__controls">
-                            <a href="{{ url('admin_category_edit', {id: categoryDetail.category.id}) }}" class="form-tree__item__control">
+                            <a href="{{ url('admin_category_edit', {id: categoryWithPreloadedChildren.category.id}) }}" class="form-tree__item__control">
                                 <i class="svg svg-pencil"></i>
                             </a>
                             {% set csrfTokenId = constant('Shopsys\\FrameworkBundle\\Component\\Router\\Security\\RouteCsrfProtector::CSRF_TOKEN_ID_PREFIX') ~ 'admin_category_delete' %}
                             {% set categoryDeleteUrl = url('admin_category_delete', {
-                                id: categoryDetail.category.id,
+                                id: categoryWithPreloadedChildren.category.id,
                                 (constant('Shopsys\\FrameworkBundle\\Component\\Router\\Security\\RouteCsrfProtector::CSRF_TOKEN_REQUEST_PARAMETER')): csrf_token(csrfTokenId)
                             }) %}
                             <a
@@ -47,7 +47,7 @@
                             </a>
                         </span>
                     </div>
-                    {{ self.categoryTreeItem(categoryDetail.children, false) }}
+                    {{ self.categoryTreeItem(categoryWithPreloadedChildren.children, false) }}
                 </li>
             {% endfor %}
         </ul>
@@ -56,7 +56,7 @@
     {% if isForAllDomains %}
         <div class="form-line">
             <div id="js-category-tree-sorting" class="form-tree form-tree--open form-tree--dragable">
-                {{ self.categoryTreeItem(categoryDetails, true) }}
+                {{ self.categoryTreeItem(categoriesWithPreloadedChildren, true) }}
             </div>
         </div>
 
@@ -75,7 +75,7 @@
     {% else %}
         <div class="form-line">
             <div class="form-tree form-tree--open form-tree--dragable">
-                {{ self.categoryTreeItem(categoryDetails, true) }}
+                {{ self.categoryTreeItem(categoriesWithPreloadedChildren, true) }}
             </div>
         </div>
         <div>
